+++ src/views/pages/admin/AdminEventCreate.tsx
@@ -1,30 +1,33 @@
+import { useCallback, useEffect, useState } from 'react';
 import { useNavigate } from 'react-router';
 import { useAppDispatch, useAppSelector } from 'hooks/hooks';
 import useEventErrorModal from 'hooks/useEventErrorModal';
 import { IEventRequest } from 'types/EventData.type';
-import { createEvent } from 'redux/eventSlice';
+import { createEvent, getEventByUserId } from 'redux/eventSlice';
+import { getAdminAuth } from 'services/auth.service';
+import { getUserById as getAdminUserById } from 'redux/adminUserSlice';
+import SessionServices from 'services/session.services';
 
 import AdminPageLayout from 'views/components/molecules/Layout/AdminPageLayout';
 import EditEventForm from 'views/components/organisms/EditEventForm';
 import ErrorModal from 'views/components/organisms/ErrorModal';
-import { useCallback, useEffect } from 'react';
 import { StateStatus } from 'types/StateStatus.type';
 
 const AdminEventCreate = () => {
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
+  const [adminUserId] = useState(SessionServices.getAdminUserId());
 
+  const { SK: userId } = useAppSelector((state) => state.adminUser.user);
   const { event, status: stateStatus } = useAppSelector((state) => state.event);
   const { event: EventStateStatus } = stateStatus;
   const { SK: eventId } = event;
 
-  const {
-    status,
-    errorMessages,
-    closeModalHandler,
-    showModalHandler,
-    isModalShown,
-  } = useEventErrorModal();
+  console.log('status:', EventStateStatus);
+  console.log('eventId:', eventId);
+
+  const { errorMessages, closeModalHandler, showModalHandler, isModalShown } =
+    useEventErrorModal();
 
   const navigateToEventDetailIfEventExist = useCallback(() => {
     if (EventStateStatus === StateStatus.fulfilled && eventId) {
@@ -46,6 +49,20 @@ const AdminEventCreate = () => {
   };
 
   useEffect(() => {
+    const token = getAdminAuth();
+
+    if (adminUserId && token) {
+      dispatch(getAdminUserById({ userId: adminUserId, token }));
+    }
+  }, [dispatch, adminUserId]);
+
+  useEffect(() => {
+    if (userId) {
+      dispatch(getEventByUserId(userId));
+    }
+  }, [userId]);
+
+  useEffect(() => {
     navigateToEventDetailIfEventExist();
   }, [navigateToEventDetailIfEventExist]);
 
@@ -59,12 +76,18 @@ const AdminEventCreate = () => {
         buttonStyle="bg-Pink-default text-white"
       />
       <AdminPageLayout>
-        <h2 className="mb-2">Create invitations</h2>
-        <EditEventForm
-          className="w-4/5"
-          updateButtonText="Create invitations"
-          formSubmitLogic={formSubmitLogic}
-        />
+        {(!EventStateStatus || EventStateStatus === StateStatus.pending) &&
+          !eventId && <div>Loading...</div>}
+        {EventStateStatus === StateStatus.fulfilled && !eventId && (
+          <>
+            <h2 className="mb-2">Create invitations</h2>
+            <EditEventForm
+              className="w-4/5"
+              updateButtonText="Create invitations"
+              formSubmitLogic={formSubmitLogic}
+            />
+          </>
+        )}
       </AdminPageLayout>
     </>
   );
